Move Issue #9 tests from unit to integration tests
Replace unit tests (Issue9CaseSensitivePropertyTests + Issue9EdgeCaseTests)
with a single integration test class that exercises the full CosmosClient →
FakeCosmosHandler → InMemoryContainer pipeline. 22 tests covering:

- JObject and strong-type round-trips with case-sensitive properties (d/D)
- Mixed casing combinations (a/A/aa/AA/aA/Aa)
- PascalCase preservation (no camelCase transformation)
- System properties (_etag, _ts, _rid)
- Enum round-trips
- Upsert/Replace operations
- SQL queries (SELECT, WHERE, nested property access)
- Patch operations (top-level and nested)
- [JsonProperty] attribute overrides
- Deep nesting and arrays with case-sensitive properties
- All 52 single-letter properties (a-z, A-Z)
- Cross-document casing independence
